🚀 Refactor storage middlewares to different folder
